refactor(tabs): update GeneralTab test setup and improve configuration saving logic
